feat: finding cycles for any graph type
