+++ src/main/kotlin/graphs/algo/ToAdjacencyList.kt
