+++ src/main/kotlin/graphs/types/DirectedGraph.kt
@@ -7,8 +7,4 @@ class DirectedGraph<V, E> : UnweightedGraph<V, E>() {
     fun findBridges(){
         TODO("Not yet implemented")
     }
-
-    override fun findCycle(v: V): List<V> {
-        TODO("Not yet implemented")
-    }
 }
