+++ src/main/kotlin/graphs/types/WeightedDirectedGraph.kt
@@ -1,6 +1,7 @@
 package graphs.types
 
 import graphs.abstracts.WeightedGraph
+import graphs.algo.toAdjacencyList
 
 class WeightedDirectedGraph<V, E>: WeightedGraph<V, E>() {
 
@@ -9,10 +10,6 @@ class WeightedDirectedGraph<V, E>: WeightedGraph<V, E>() {
     }
 
     override fun MinimalPathFloidBellman(start: V, target: V): WeightedGraph<V, E> {
-        TODO("Not yet implemented")
-    }
-
-    override fun findCycle(v: V): List<V> {
         TODO("Not yet implemented")
     }
 }
